서버에서 좀비 경로에 따라 이동 계산을 deltatime을 이용하여 계산, 10주차 작업일지

+++ unreal/test/Server/Game Server/Server/iocpServerClass.h
@@ -150,7 +150,7 @@ private:
 
 	ZombieController* zombieclass;
 
-	int GameTime = 0;
+	float GameTime = 0.f;
 
 	bool b_Timer = false;
 };
